# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M3th",
        "description": "math testing",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://node43.webte.fei.stuba.sk",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/tests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get a list of tests (optionally filter by title)",
                "operationId": "c111461571e8c1952d5ec29f6b79e9ce",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by test title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tests": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Test"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Create a new test with question IDs",
                "operationId": "ce2fcda95b04246f42b319d2f2dce318",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "question_ids"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Math Test"
                                    },
                                    "question_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Test created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tests/{test}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get a single test with its questions",
                "operationId": "0da1154a29f20a65736b20deb2592430",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update a test's title and question IDs",
                "operationId": "8b6bfbf28a984df90e7144ef9409eb1f",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "question_ids"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Math Test"
                                    },
                                    "question_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete a test and detach its questions",
                "operationId": "2d92c6746348690a46bb2884777a2b13",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Test deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Test": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "type": "object"
            },
            "Question": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "assignment_sk": {
                        "type": "string"
                    },
                    "assignment_en": {
                        "type": "string"
                    },
                    "isMultiChoice": {
                        "type": "boolean"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
            "Answer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "question_id": {
                        "type": "integer"
                    },
                    "answer_sk": {
                        "type": "string"
                    },
                    "answer_en": {
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name_sk": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}